+++ src/main/java/blackjack/domain/deck/CardShuffleStrategy.java
@@ -1,5 +1,6 @@
-package blackjack.model;
+package blackjack.domain.deck;
 
+import blackjack.domain.card.Card;
 import java.util.List;
 
 public interface CardShuffleStrategy {
