+++ src/main/java/blackjack/domain/deck/RandomShuffleStrategy.java
@@ -1,5 +1,6 @@
-package blackjack.model;
+package blackjack.domain.deck;
 
+import blackjack.domain.card.Card;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
