+++ src/main/java/blackjack/domain/result/GameResult.java
@@ -1,4 +1,6 @@
-package blackjack.model;
+package blackjack.domain.result;
+
+import blackjack.domain.participant.User;
 
 public record GameResult(String name, int profit) {
     public static GameResult from(User user, int profit) {
