+++ src/main/java/blackjack/domain/result/GameSummary.java
@@ -1,5 +1,7 @@
-package blackjack.model;
+package blackjack.domain.result;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.participant.User;
 import java.util.List;
 
 public record GameSummary(String name, List<Card> cards, int score, boolean isBust, boolean isBlackjack) {
