+++ src/test/java/blackjack/domain/deck/FixShuffleStrategy.java
@@ -1,5 +1,6 @@
-package blackjack.model;
+package blackjack.domain.deck;
 
+import blackjack.domain.card.Card;
 import java.util.ArrayList;
 import java.util.List;
