+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -1,8 +1,12 @@
-package blackjack.model;
+package blackjack.domain.participant;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Suit;
+import blackjack.domain.result.GameOutcome;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
