+++ src/test/java/blackjack/domain/participant/UserTest.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
